+++ src/components/App/MobileTabBar/index.js
@@ -4,10 +4,10 @@ import PropTypes from 'prop-types';
 import cn from 'classnames';
 
 import styles from './MobileTabBar.module.css';
-import inboxButtonStyles from '../Header/InboxButton.module.css';
+import inboxIconStyles from '../Header/InboxIcon.module.css';
 import Magnifiner from 'common/icons/Magnifiner';
 import PlusCircle from 'common/icons/PlusCircle';
-import { InboxButtonIcon } from '../Header/InboxButton';
+import InboxIcon from '../Header/InboxIcon';
 
 const ShareButton = ({ className, ...props }) => {
   return <PlusCircle className={cn(styles.share, className)} {...props} />;
@@ -25,7 +25,7 @@ const MobileTabBar = ({ focusSearch }) => {
   return (
     <nav className={styles.tabBar}>
       <Link to="/notifications" className={styles.tabItem}>
-        <InboxButtonIcon className={inboxButtonStyles.tabBarIcon} />
+        <InboxIcon className={inboxIconStyles.tabBarIcon} />
         <span className={styles.label}>通知</span>
       </Link>
       <Link to="/share" className={styles.tabItem}>
